Chapter 23

+++ controllers/admin.js
@@ -1,5 +1,7 @@
 const Product = require('../models/product');
+const fileHelper = require('../util/file');
 const { validationResult } = require('express-validator');
+
 exports.getAddProduct = (req, res, next) => {
   
   res.render('admin/edit-product', {
@@ -20,8 +22,8 @@ exports.postAddProduct = (req, res, next) => {
   const errors = validationResult(req);
 
   if (!errors.isEmpty()) {
-    return res.status(422)
-    .render('admin/edit-product', {
+    console.log(errors.array());
+    return res.status(422).render('admin/edit-product', {
       pageTitle: 'Add Product',
       path: '/admin/add-product',
       editing: false,
@@ -36,6 +38,7 @@ exports.postAddProduct = (req, res, next) => {
       validationErrors: errors.array()
     });
   }
+
   const product = new Product({
     title: title,
     price: price,
@@ -46,7 +49,6 @@ exports.postAddProduct = (req, res, next) => {
   product
     .save()
     .then(result => {
-      // console.log(result);
       console.log('Created Product');
       res.redirect('/admin/products');
     })
@@ -54,23 +56,6 @@ exports.postAddProduct = (req, res, next) => {
       const error = new Error(err);
       error.httpStatusCode = 500;
       return next(error);
-      
-      // res.redirect('/500');
-      // return res.status(500)
-      //   .render('admin/edit-product', {
-      //     pageTitle: 'Add Product',
-      //     path: '/admin/add-product',
-      //     editing: false,
-      //     hasError: true,
-      //     product: {
-      //       title: title,
-      //       imageUrl: imageUrl,
-      //       price: price,
-      //       description: description
-      //     },
-      //     errorMessage: 'Database operation failed, please try again.',
-      //     validationErrors: []
-      //   });
     });
 };
 
@@ -106,7 +91,7 @@ exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
   const updatedPrice = req.body.price;
-  const updatedImageUrl = req.body.imageUrl;
+  const image = req.file;
   const updatedDesc = req.body.description;
   const errors = validationResult(req);
 
@@ -119,7 +104,6 @@ exports.postEditProduct = (req, res, next) => {
       hasError: true,
       product: {
         title: updatedTitle,
-        imageUrl: updatedImageUrl,
         price: updatedPrice,
         description: updatedDesc,
         _id: prodId
@@ -136,7 +120,10 @@ exports.postEditProduct = (req, res, next) => {
       product.title = updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDesc;
-      product.imageUrl = updatedImageUrl;
+
+      if (image) {
+        product.imageUrl = image.path;
+      }
       return product.save()
       .then(result => {
         console.log('UPDATED PRODUCT!');
@@ -152,10 +139,7 @@ exports.postEditProduct = (req, res, next) => {
 
 exports.getProducts = (req, res, next) => {
   Product.find({userId: req.user._id})
-    // .select('title price -_id')
-    // .populate('userId', 'name')
     .then(products => {
-      console.log(products);
       res.render('admin/products', {
         prods: products,
         pageTitle: 'Admin Products',
@@ -169,16 +153,20 @@ exports.getProducts = (req, res, next) => {
     });
 };
 
-exports.postDeleteProduct = (req, res, next) => {
-  const prodId = req.body.productId;
-  Product.deleteOne({_id: prodId, userId: req.user._id})
-    .then(() => {
-      console.log('DESTROYED PRODUCT');
-      res.redirect('/admin/products');
-    })
-    .catch(err => {
-      const error = new Error(err);
-      error.httpStatusCode = 500;
-      return next(error);
-    });
+exports.deleteProduct = (req, res, next) => {
+  const prodId = req.params.productId;
+  Product.findById(prodId)
+  .then(product => {
+    if (!product) {
+      return next(new Error('Product not found!'));
+    }
+    return Product.deleteOne({_id: prodId, userId: req.user._id});   
+  })
+  .then(() => {
+    console.log('DESTROYED PRODUCT');
+    res.status(200).json({message: 'Success!'});
+  })
+  .catch(err => {
+    res.status(500).json({message: 'Deleting product failed!'})
+  }); 
 };
